snap4flame-backend/pyproject.toml
[project]
name = "snap4frame-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Panaintescu Adrian Valentin", email = "adrian.cert@gmail.com"},
]
dependencies = [
    "aiohttp",
    "python-dotenv",
    "aiohttp-session",
    "google-cloud-pubsub",
    "google-cloud-ndb",
    "websockets>=12.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[tool.ruff]
extend-select = ["C", "N", "SIM", "PTH", "I", "PL", "FURB"]
